pyproject.toml
[tool.poetry]
name = "ftp"
version = "0.1.0"
description = ""
authors = ["Ivan Chekunkov <chekunkov.ivan@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12, <3.13"
loguru = "^0.7.2"
pyinstaller = "^6.7.0"
ruff = "^0.5.4"
isort = "^5.13.2"
pyyaml = "^6.0.2"
aioftp = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"